FMOD Changes
note: the event Player_fire_throw has 2 parameters: which_fire_loops (goes from 1 to 3 to set which fire throw SFX will loop). I tried to make a random modulation so the values would automatically set but it's not working so its up to the programer to set the randomness in that.
the other parameter is fire_throw_stops and its value can be 0 or 1. 0 is when the fire throw loop is still playing (the player is pressing the mouse for the fire throw) and 1 its when the player releases the mouse